project/requirements.txt
torch>=2.1.0
torchvision>=0.17.0
transformers>=4.50.0
datasets>=2.13.0
mlflow>=2.7.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.2
Pillow>=10.0.0
gradio>=3.50.0

